feat: update properties
